+++ app/src/main/java/com/styudint/nepyatnashki/data/AndroidGameState.kt
@@ -3,5 +3,15 @@ package com.styudint.nepyatnashki.data
 import androidx.lifecycle.LiveData
 
 interface AndroidGameState : GameState {
+    fun handleTap(value: Int)
     fun stopWatch(): LiveData<Long>
+
+    fun startTime(): Long
+    fun gameTime(): Long
+
+    fun subscribe(listener: AndroidGameStateListener)
+    fun unsubscribe(listener: AndroidGameStateListener)
+
+    fun start()
+    fun stop()
 }
